uniform some typemaps + fix warning with python.h inclusion + implement draw method in coreimage

+++ pymtcore/src/texture.h
@@ -40,7 +40,7 @@ public:
 
 	unsigned int	id;
 	unsigned int	target;
-	double			tex_coords[12];
+	float			tex_coords[12];
 
 protected:
 	virtual ~Texture();
